Revert "Refactor: rename the folder name utility to utilities"
