+++ UniMolXC/network/kernel/_xcnet.py
@@ -4,7 +4,7 @@
 from torch import nn
 
 # local modules
-from UniMolXC.network.utilities.xcloss import tminnesota
+from UniMolXC.network.utility.xcloss import tminnesota
 
 class XCParameterizationNet(nn.Module):
     '''
